chore: consolidating keypads and removing duplicate
