+++ app/components/UI/Ramp/Deposit/Views/BuildQuote/BuildQuote.tsx
@@ -609,7 +609,6 @@ const BuildQuote = () => {
             onChange={handleKeypadChange}
             currency={fiatCurrency.symbol}
             decimals={0}
-            deleteIcon={<Icon name={IconName.Arrow2Left} size={IconSize.Lg} />}
           />
         </ScreenLayout.Content>
       </ScreenLayout.Body>
